Move oyez_brief.py into algorithms module
